+++ src/main/java/de/marhali/easyi18n/assistance/completion/JavaCompletionContributor.java
@@ -3,14 +3,15 @@ package de.marhali.easyi18n.assistance.completion;
 import com.intellij.codeInsight.completion.CompletionContributor;
 import com.intellij.codeInsight.completion.CompletionType;
 import com.intellij.patterns.PlatformPatterns;
-import com.intellij.psi.PsiLiteralValue;
+import com.intellij.psi.PsiLiteralExpression;
 
 /**
- * Java specific completion contributor
+ * Java specific completion contributor.
  * @author marhali
  */
 public class JavaCompletionContributor extends CompletionContributor {
     public JavaCompletionContributor() {
-        extend(CompletionType.BASIC, PlatformPatterns.psiElement().inside(PsiLiteralValue.class), new KeyCompletionProvider());
+        extend(CompletionType.BASIC, PlatformPatterns.psiElement().inside(PsiLiteralExpression.class),
+                new KeyCompletionProvider());
     }
 }