+++ src/main/java/de/marhali/easyi18n/assistance/folding/JsFoldingBuilder.java
@@ -1,9 +1,9 @@
 package de.marhali.easyi18n.assistance.folding;
 
 import com.intellij.lang.ASTNode;
+import com.intellij.lang.javascript.psi.JSLiteralExpression;
 import com.intellij.openapi.util.Pair;
 import com.intellij.psi.PsiElement;
-import com.intellij.psi.PsiLiteralExpression;
 import com.intellij.psi.util.PsiTreeUtil;
 
 import org.jetbrains.annotations.NotNull;
@@ -13,20 +13,20 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 /**
- * Java specific translation key folding.
+ * JavaScript specific translation key folding.
  * @author marhali
  */
-public class JavaFoldingBuilder extends AbstractFoldingBuilder {
+public class JsFoldingBuilder extends AbstractFoldingBuilder {
     @Override
     @NotNull List<Pair<String, PsiElement>> extractRegions(@NotNull PsiElement root) {
-        return PsiTreeUtil.findChildrenOfType(root, PsiLiteralExpression.class).stream().map(literalExpression ->
-                        Pair.pair(String.valueOf(literalExpression.getValue()), (PsiElement) literalExpression))
+        return PsiTreeUtil.findChildrenOfType(root, JSLiteralExpression.class).stream().map(literalExpression ->
+                        Pair.pair(literalExpression.getStringValue(), (PsiElement) literalExpression))
                 .collect(Collectors.toList());
     }
 
     @Override
     @Nullable String extractText(@NotNull ASTNode node) {
-        PsiLiteralExpression literalExpression = node.getPsi(PsiLiteralExpression.class);
-        return String.valueOf(literalExpression.getValue());
+        JSLiteralExpression literalExpression = node.getPsi(JSLiteralExpression.class);
+        return literalExpression.getStringValue();
     }
 }